% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_initial_par}
\alias{data_initial_par}
\title{Initialise model parameters}
\usage{
data_initial_par(data)
}
\arguments{
\item{data}{A data frame containing counts `n` and `observed_mu`.}
}
\value{
A data frame with an added `start_par` column.
}
\description{
Provides starting values for latent parameters based on counts.
}
