% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{exponential_kernel}
\alias{exponential_kernel}
\title{Exponential kernel function}
\usage{
exponential_kernel(distance, sigma)
}
\arguments{
\item{sigma}{The length scale parameter, controlling the rate of decay}

\item{x}{A scalar representing the first input point}

\item{x_prime}{A scalar or vector representing the second input point(s)}
}
\description{
Exponential kernel function
}
