% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{get_temporal_distance}
\alias{get_temporal_distance}
\title{Pairwise temporal distances}
\usage{
get_temporal_distance(times)
}
\arguments{
\item{times}{A numeric vector of time indices.}
}
\value{
A symmetric matrix of pairwise temporal distances.
}
\description{
Computes pairwise distances between time points.
}
