% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{gp_build_state}
\alias{gp_build_state}
\title{Build state object}
\usage{
gp_build_state(obs_data, coordinates, hyperparameters, n, nt, period = 52)
}
\arguments{
\item{obs_data}{Observation data}

\item{coordinates}{Site coordinates}

\item{hyperparameters}{Vector of hyperparameters}

\item{n}{N sites}

\item{nt}{N times}

\item{period}{Period}
}
\value{
List of pre-computaed GP elements
}
\description{
Build state object
}
