% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{observed_data}
\alias{observed_data}
\title{Observed data generation}
\usage{
observed_data(data, p_one, p_switch)
}
\arguments{
\item{data}{Output of [simulate_data()].}

\item{p_one}{Probability that a new missingness cluster starts with a missing
value.}

\item{p_switch}{Probability of switching between missing and observed
clusters.}
}
\value{
Tibble with observed counts `y_obs` and inferred `mu_infer`,
  `z_infer` and `f_infer`.
}
\description{
Observed data generation
}
\details{
Missingness is generated stochastically via
  [generate_clustered_binary()]. Set a seed with [set.seed()] to reproduce
  results.
}
