% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_order_index}
\alias{data_order_index}
\title{Order data and assign identifiers}
\usage{
data_order_index(data, ...)
}
\arguments{
\item{data}{A data frame containing site identifiers and time `t`.}

\item{...}{Columns identifying sites passed to [dplyr::group_by()]
(unquoted).}
}
\value{
A data frame ordered by site and time with an `id` column.
}
\description{
Arranges data by site and time and creates a factor `id` per site.
}
