% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hyperparameters.R
\name{fit}
\alias{fit}
\title{Fit hyperparmeters}
\usage{
fit(
  obs_data,
  nt,
  period,
  n_sites,
  mask_prop = 0.2,
  verbose = FALSE,
  par0 = c(1, 5, 100),
  lower = c(1e-04, 0.8, 52 * 1.5),
  upper = c(2, 10, 500)
)
}
\description{
Fit hyperparmeters
}
