% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{get_spatial_distance}
\alias{get_spatial_distance}
\title{Pairwise spatial distances}
\usage{
get_spatial_distance(coordinates)
}
\arguments{
\item{coordinates}{A data frame with columns `lon` and `lat` in degrees.}
}
\value{
A symmetric matrix of pairwise spatial distances.
}
\description{
Computes pairwise Euclidean distances between locations.
}
