% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{kdiag_from_factors}
\alias{kdiag_from_factors}
\title{Kronecker diadiagonalg}
\usage{
kdiag_from_factors(space_diag, time_diag, n, nt)
}
\arguments{
\item{space_diag}{Space matrix diagonal}

\item{time_diag}{Time matrix diagonal}

\item{n}{Number of sites}

\item{nt}{Number of times}
}
\value{
Kronecker diagonal
}
\description{
A helper: Kronecker diagonal without allocating a big dense kronecker()
}
