% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kron.R
\name{kron_mv}
\alias{kron_mv}
\title{Fast Kronecker-product matrix-vector multiply (times vary fastest)}
\usage{
kron_mv(v, space, time)
}
\arguments{
\item{v}{Numeric vector of length nrow(space) * nrow(time).}

\item{space}{Spatial kernel matrix (n x n).}

\item{time}{Temporal kernel matrix (nt x nt).}
}
\value{
Numeric vector of the same length as v.
}
\description{
In plain terms: multiplies a big covariance K = K_space (x) K_time by a
vector without ever forming K, using a reshape-multiply-reshape trick.
}
\details{
Technically: for v = vec(t(X)) with times varying fastest,
  (K_space (x) K_time) v = vec(t(K_space %*% X %*% t(K_time)))
}
