% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{quick_mvnorm}
\alias{quick_mvnorm}
\title{Quick multivariate normal samples over two dimensions}
\usage{
quick_mvnorm(space, time)
}
\arguments{
\item{space}{Space kernel matrix}

\item{time}{Time kernel matrix}
}
\description{
Quick multivariate normal samples over two dimensions
}
