% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{quick_mvnorm}
\alias{quick_mvnorm}
\title{Quick multivariate normal samples over two dimensions}
\usage{
quick_mvnorm(space, time)
}
\arguments{
\item{space}{Space kernel matrix}

\item{time}{Time kernel matrix}
}
\description{
This is equivalent to estimating the full spatio-temporal covariance matrix
and sampling from the multivariate normal distribution:
full_k <- kronecker(dist_k, time_k)
f  <- mvrnorm(1, rep(0, n * nt), full_k)
}
