% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{quick_mvnorm_chol}
\alias{quick_mvnorm_chol}
\title{Quick multivariate normal samples over two dimensions (cholesky precomputed)}
\usage{
quick_mvnorm_chol(space_chol, time_chol)
}
\arguments{
\item{space_chol}{Cholesky decomposition of sapace kernel matrix}

\item{time_chol}{Cholesky decomposition of time kernel matrix}
}
\description{
This is equivalent to estimating the full spatio-temporal covariance matrix
and sampling from the multivariate normal distribution:
full_k <- kronecker(dist_k, time_k)
f  <- mvrnorm(1, rep(0, n * nt), full_k)
}
