% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{rbf_kernel}
\alias{rbf_kernel}
\title{RBF kernel function}
\usage{
rbf_kernel(x, x_prime, sigma)
}
\arguments{
\item{x}{A scalar representing the first input point}

\item{x_prime}{A scalar or vector representing the second input point(s)}

\item{sigma}{The length scale parameter, controlling the smoothness of the kernel}
}
\description{
RBF kernel function
}
