% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{rbf_kernel}
\alias{rbf_kernel}
\title{Radial basis function kernel}
\usage{
rbf_kernel(x, theta)
}
\arguments{
\item{x}{A numeric vector or matrix of distances.}

\item{theta}{A positive numeric scalar giving the length-scale parameter.}
}
\value{
A numeric vector or matrix with RBF kernel values.
}
\description{
Computes the radial basis function (RBF) kernel for a distance vector
or matrix.
}
