% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.R
\name{time_kernel}
\alias{time_kernel}
\title{Extimate the temporal kernel}
\usage{
time_kernel(
  times,
  periodic_scale,
  long_term_scale,
  nugget = 1e-09,
  period = 52
)
}
\description{
Extimate the temporal kernel
}
